KE-13969 record license capacity from double to long

+++ src/core-metadata/src/main/java/io/kyligence/kap/metadata/sourceusage/LicenseInfo.java
@@ -45,7 +45,7 @@ public class LicenseInfo implements Serializable {
     private long currentCapacity = 0L;
 
     @JsonProperty("capacity")
-    private double capacity = 0L;
+    private long capacity = 0L;
 
     @JsonProperty("capacity_status")
     private CapacityStatus capacityStatus = CapacityStatus.OK;
